完成了Lab2

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,8 +5,8 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) {
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
@@ -26,5 +26,6 @@ public class DebugExercise1 {
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
+
     }
 }
